requirements.txt
# mysql-connector-python
pandas
apache-airflow-providers-http
apache-airflow-providers-mysql